+++ src/Instance/ViewInstance/ViewInstance.js
@@ -88,6 +88,7 @@ const ViewInstance = (props) => {
   const history = useHistory();
   const location = useLocation();
   const paneTitleRef = useRef(null);
+  const accordionStatusRef = useRef();
 
   const centralTenantId = stripes.user.user?.consortium?.centralTenantId;
 
@@ -223,6 +224,8 @@ const ViewInstance = (props) => {
     marcRecord,
     callout,
     canBeOpenedInLinkedData,
+    accordionStatusRef,
+    onCopy,
   });
 
   const mutateInstance = async (entity, { onError }) => {
@@ -332,6 +335,7 @@ const ViewInstance = (props) => {
         holdingsSection={holdingsSection}
         paneTitleRef={paneTitleRef}
         userTenantPermissions={userTenantPermissions}
+        accordionStatusRef={accordionStatusRef}
       />
       <InstanceModals
         instance={instance}